Move console config code to separate class
